# lib/api-zod/src/generated/types/attendee.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Attendee {
  id: number;
  qrId: string;
  name: string;
  /** @nullable */
  email?: string | null;
  ticketType: string;
  /** @nullable */
  checkedInAt: Date | null;
  createdAt: Date;
}