# lib/api-zod/src/generated/types/attendeeImportInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendeeInput } from './attendeeInput';

export interface AttendeeImportInput {
  /** @minItems 1 */
  attendees: AttendeeInput[];
}