# lib/api-zod/src/generated/types/attendeeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttendeeInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  /** @minLength 1 */
  ticketType: string;
  qrId?: string;
}