# lib/api-zod/src/generated/types/generateInvitationsInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateInvitationsInput {
  /** @minimum 0 */
  vipCount: number;
  /** @minimum 0 */
  regularCount: number;
}